"instant" game resolution for both player's and/or dealer's blackjack

+++ src/com/ivanfoong/insightdataengineering/blackjack/user/Person.java
@@ -9,7 +9,7 @@ public class Person {
     private Wallet mWallet;
 
     public Person(final Integer aTotalValue) {
-        mWallet = new Wallet(aTotalValue);
+        mWallet = new Wallet(aTotalValue.doubleValue());
     }
 
     public Wallet getWallet() {
