+++ src/com/ivanfoong/insightdataengineering/blackjack/wallet/Wallet.java
@@ -4,21 +4,21 @@ package com.ivanfoong.insightdataengineering.blackjack.wallet;
  * Created by ivanfoong on 4/4/14.
  */
 public class Wallet {
-    private Integer mTotalValue;
+    private Double mTotalValue;
 
-    public Wallet(final Integer aTotalValue) {
+    public Wallet(final Double aTotalValue) {
         mTotalValue = aTotalValue;
     }
 
-    public Integer getTotalValue() {
+    public Double getTotalValue() {
         return mTotalValue;
     }
 
-    public void increaseValue(final Integer aValue) {
+    public void increaseValue(final Double aValue) {
         mTotalValue += aValue;
     }
 
-    public void decreaseValue(final Integer aValue) throws NegativeWalletValueException {
+    public void decreaseValue(final Double aValue) throws NegativeWalletValueException {
         if ((mTotalValue - aValue) < 0) {
             throw new NegativeWalletValueException();
         }
